requirements.txt
django
psycopg2-binary
django-environ
django-cors-headers
langchain
langchain-community
langchain-postgres
langchain-google-genai
pypdf
pgvector
python-dotenv
djangorestframework
google-auth
django-storages
google-cloud-storage
google-cloud-documentai
google-generativeai
dj-database-url
whitenoise
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dj-rest-auth
django-allauth
djangorestframework-simplejwt
cryptography
gunicorn
